requirements.txt
streamlit
langchain
langchain-openai
langchain-core
openai
python-dotenv
